+++ src/accuraterip.rs
@@ -415,6 +415,10 @@ impl AccurateRip {
 	///     toc.accuraterip_id().pretty_print(),
 	///     "013-0015deca-00d9b921-9a0a6e0d",
 	/// );
+	/// assert_eq!(
+	///     toc.accuraterip_id().to_string(),
+	///     "013-0015deca-00d9b921-9a0a6e0d",
+	/// );
 	/// ```
 	pub fn pretty_print(&self) -> String {
 		let mut out: Vec<u8> = vec![
@@ -548,6 +552,7 @@ mod tests {
 			let toc = Toc::from_cdtoc(t).expect("Invalid TOC");
 			let ar_id = toc.accuraterip_id();
 			assert_eq!(ar_id.to_string(), id);
+			assert_eq!(ar_id.pretty_print(), id);
 
 			// Test decoding three ways.
 			assert_eq!(AccurateRip::decode(id), Ok(ar_id));